requirements.txt
requests==2.32.3
PyYAML==6.0.2
Jinja2==3.1.4
deepdiff==7.0.1
tqdm==4.66.4

